requirements.txt
fastapi
uvicorn
langchain
langchain-openai
chromadb
pydantic
python-dotenv
fhir.resources